requirements.txt
# Core Framework
streamlit>=1.28.0

# NLP and Machine Learning - Compatible versions
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
spacy>=3.6.0
nltk>=3.8.1

# Document Processing
PyMuPDF>=1.23.0
python-docx>=0.8.11
Pillow>=10.0.0

# Data Processing - Fixed versions for compatibility
pandas>=2.0.0,<2.3.0
numpy>=1.24.0,<2.0.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0

# Database
sqlalchemy>=2.0.0

# Security and Authentication
bcrypt>=4.0.0
python-dotenv>=1.0.0

# Specific compatible versions to avoid conflicts
gensim>=4.3.0,<4.4.0
numba>=0.58.0,<0.61.0
pyasn1>=0.4.6,<0.5.0
pyasn1-modules>=0.2.8,<0.3.0